refactor: separate component into view and edit
