+++ src/blocks/download/block.json
@@ -2,7 +2,6 @@
 	"$schema": "https://schemas.wp.org/trunk/block.json",
 	"apiVersion": 2,
 	"name": "inseri-core/download",
-	"version": "0.1.0",
 	"title": "Download",
 	"category": "inseri",
 	"description": "Let visitor download block data",
