+++ src/blocks/dropdown/index.tsx
@@ -1,12 +1,11 @@
 import { IconCaretDown } from '@tabler/icons-react'
 import { useBlockProps } from '@wordpress/block-editor'
-import type { BlockEditProps, BlockSaveProps } from '@wordpress/blocks'
+import type { BlockSaveProps } from '@wordpress/blocks'
 import { registerBlockType } from '@wordpress/blocks'
 import stringify from 'json-stable-stringify'
-import { SetupEditorEnv } from '../../components'
 import { ConsumerBeacon, ProducerBeacon } from '../../globalScript'
 import json from './block.json'
-import { DropdownEdit } from './Component'
+import Edit from './edit'
 
 const { name, ...settings } = json as any
 
@@ -20,25 +19,15 @@ export interface Attributes {
 	clearable: boolean
 }
 
-function Edit(props: BlockEditProps<Attributes>) {
-	return (
-		<SetupEditorEnv {...props} baseBlockName={'dropdown'}>
-			<DropdownEdit {...props} />
-		</SetupEditorEnv>
-	)
-}
-
-function Save({ attributes }: BlockSaveProps<Attributes>) {
-	return (
-		<div {...useBlockProps.save()} data-attributes={stringify(attributes)}>
-			is loading ...
-		</div>
-	)
-}
-
 registerBlockType<Attributes>(name, {
 	...settings,
 	edit: Edit,
-	save: Save,
+	save: ({ attributes }: BlockSaveProps<Attributes>) => {
+		return (
+			<div {...useBlockProps.save()} data-attributes={stringify(attributes)}>
+				is loading ...
+			</div>
+		)
+	},
 	icon: <IconCaretDown style={{ fill: 'none' }} />,
 })
