+++ src/blocks/html/block.json
@@ -2,7 +2,6 @@
 	"$schema": "https://schemas.wp.org/trunk/block.json",
 	"apiVersion": 2,
 	"name": "inseri-core/html",
-	"version": "0.1.0",
 	"title": "HTML Code",
 	"category": "inseri",
 	"description": "Render HTML code",
