+++ src/blocks/html/edit.tsx
@@ -1,16 +1,18 @@
 import { useAvailableBeacons, useWatch } from '@inseri/lighthouse'
-import { IconHtml, IconCircleOff } from '@tabler/icons-react'
+import { IconHtml } from '@tabler/icons-react'
 import { BlockControls, InspectorControls } from '@wordpress/block-editor'
 import type { BlockEditProps } from '@wordpress/blocks'
 import { PanelBody, PanelRow, TextControl, ToolbarButton, ToolbarGroup } from '@wordpress/components'
 import { useEffect } from '@wordpress/element'
 import { __ } from '@wordpress/i18n'
 import { edit } from '@wordpress/icons'
-import { Box, CodeEditor, createStyles, Group, Select, Text, useGlobalState } from '../../components'
+import { Box, Group, Select, SetupEditorEnv, StateProvider, Text, useGlobalState } from '../../components'
+import json from './block.json'
 import { Attributes } from './index'
-import { GlobalState } from './state'
+import { GlobalState, storeCreator } from './state'
+import View from './view'
 
-export function HtmlEdit(props: BlockEditProps<Attributes>) {
+function EditComponent(props: BlockEditProps<Attributes>) {
 	const { isSelected } = props
 
 	const { input, blockId, blockName, isWizardMode, mode, actions } = useGlobalState((state: GlobalState) => state)
@@ -87,66 +89,19 @@ export function HtmlEdit(props: BlockEditProps<Attributes>) {
 					/>
 				</Box>
 			) : (
-				<HtmlView isGutenbergEditor isSelected={isSelected} />
+				<View isGutenbergEditor isSelected={isSelected} />
 			)}
 		</>
 	)
 }
 
-const useViewStyles = createStyles({
-	wrapper: {
-		'&:hover': {
-			borderRadius: '1px',
-			boxShadow: '0 0 0 var(--wp-admin-border-width-focus) var(--wp-admin-theme-color)',
-		},
-	},
-})
-
-interface ViewProps {
-	isGutenbergEditor?: boolean
-	isSelected?: boolean
-}
-
-export function HtmlView({ isGutenbergEditor, isSelected }: ViewProps) {
-	const { input, mode } = useGlobalState((state: GlobalState) => state)
-	const { value, status } = useWatch(input)
-	const { wrapper } = useViewStyles().classes
-
-	const isEmpty = !value || (typeof value === 'string' && !value.trim())
-	let altText = __('No HTML code is set', 'inser-core')
-
-	let preparedValue = value
-	const hasError = status === 'error' || status === 'unavailable'
-
-	if (hasError) {
-		preparedValue = ''
-		altText = __('An error has occurred', 'inser-core')
-	}
-
-	return mode === 'code' ? (
-		<Box p="md">
-			<CodeEditor type={'html'} value={preparedValue} />
-		</Box>
-	) : isEmpty || hasError ? (
-		<Group
-			align="center"
-			position="center"
-			style={{
-				background: '#F8F9FA',
-				color: '#868E96',
-				padding: '8px',
-			}}
-		>
-			<IconCircleOff size={40} />
-			<Text size="xl" align="center">
-				{altText}
-			</Text>
-		</Group>
-	) : (
-		<div
-			className={isGutenbergEditor && !isSelected ? wrapper : undefined}
-			style={{ minHeight: isGutenbergEditor ? '50px' : undefined }}
-			dangerouslySetInnerHTML={{ __html: preparedValue }}
-		/>
+export default function Edit(props: BlockEditProps<Attributes>) {
+	const { setAttributes, attributes } = props
+	return (
+		<SetupEditorEnv {...props} baseBlockName={'html'}>
+			<StateProvider stateCreator={storeCreator} keysToSave={Object.keys(json.attributes)} setAttributes={setAttributes} initialState={attributes}>
+				<EditComponent {...props} />
+			</StateProvider>
+		</SetupEditorEnv>
 	)
 }
