+++ src/blocks/image/block.json
@@ -2,7 +2,6 @@
 	"$schema": "https://schemas.wp.org/trunk/block.json",
 	"apiVersion": 2,
 	"name": "inseri-core/image",
-	"version": "0.1.0",
 	"title": "Image Box",
 	"category": "inseri",
 	"description": "Display image",
