+++ src/blocks/image/edit.tsx
@@ -1,14 +1,16 @@
 import { useAvailableBeacons, useJsonBeacons, useWatch } from '@inseri/lighthouse'
-import { IconPhoto, IconPhotoOff, IconCircleOff } from '@tabler/icons-react'
+import { IconPhoto } from '@tabler/icons-react'
 import { BlockControls, InspectorControls } from '@wordpress/block-editor'
 import type { BlockEditProps } from '@wordpress/blocks'
 import { PanelBody, PanelRow, ResizableBox, SelectControl, TextControl, ToolbarButton, ToolbarGroup } from '@wordpress/components'
 import { useEffect, useRef } from '@wordpress/element'
 import { __ } from '@wordpress/i18n'
 import { edit } from '@wordpress/icons'
-import { Box, Group, Select, Text, useGlobalState } from '../../components'
+import { Box, Group, Select, SetupEditorEnv, StateProvider, Text, useGlobalState } from '../../components'
+import json from './block.json'
 import { Attributes } from './index'
-import { GlobalState } from './state'
+import { GlobalState, storeCreator } from './state'
+import View from './view'
 
 const defaultInput = {
 	key: '',
@@ -38,7 +40,7 @@ const resizingHelps = {
 	none: __('Display image always at its original size', 'inseri-core'),
 }
 
-export function PhotoEdit(props: BlockEditProps<Attributes>) {
+function EditComponent(props: BlockEditProps<Attributes>) {
 	const { isSelected } = props
 	const { blockName, input, isWizardMode, actions, altText, caption, height, fit } = useGlobalState((state: GlobalState) => state)
 	const { updateState } = actions
@@ -151,148 +153,19 @@ export function PhotoEdit(props: BlockEditProps<Attributes>) {
 					/>
 				</Box>
 			) : (
-				<PhotoView renderResizable={renderResizable} imageRef={imageRef} isSelected={isSelected} />
+				<View renderResizable={renderResizable} imageRef={imageRef} isSelected={isSelected} />
 			)}
 		</>
 	)
 }
 
-interface ViewProps {
-	renderResizable?: (Component: JSX.Element) => JSX.Element
-	imageRef?: React.Ref<HTMLImageElement>
-	isSelected?: boolean
-}
-
-export function PhotoView({ renderResizable, imageRef, isSelected }: ViewProps) {
-	const { input, isBlob, isPrevBlob, imageUrl, prevImageUrl, caption, altText, height, fit, hasError } = useGlobalState((state: GlobalState) => state)
-	const { updateState } = useGlobalState((state: GlobalState) => state.actions)
-	const { value, status, contentType } = useWatch(input)
-
-	const isUrlEmpty = !imageUrl
-
-	const update = (url: string, blob: boolean) =>
-		updateState({
-			imageUrl: url,
-			prevImageUrl: imageUrl,
-			isBlob: blob,
-			isPrevBlob: isBlob,
-		})
-
-	useEffect(() => {
-		let processedValue = value
-
-		if (status === 'ready') {
-			if (typeof processedValue === 'string' && contentType.includes('image/svg+xml')) {
-				processedValue = new Blob([value], { type: 'image/svg+xml' })
-			}
-
-			if (typeof processedValue === 'string') {
-				update(processedValue, false)
-			}
-
-			if (typeof processedValue === 'object' && processedValue instanceof Blob) {
-				update(URL.createObjectURL(processedValue), true)
-			}
-
-			if (!processedValue) {
-				update('', false)
-			}
-
-			if (isPrevBlob) {
-				URL.revokeObjectURL(prevImageUrl)
-			}
-
-			updateState({ hasError: false })
-		}
-	}, [value])
-
-	const highlightBg = isSelected
-		? {
-				backgroundColor: '#fff',
-				backgroundImage: `repeating-linear-gradient(45deg, #808080 25%, transparent 25%, transparent 75%, #808080 75%, #808080),
-					 repeating-linear-gradient(45deg, #808080 25%, #ffffff 25%, #ffffff 75%, #808080 75%, #808080)`,
-				backgroundPosition: '0 0, 10px 10px',
-				backgroundSize: '20px 20px',
-		  }
-		: {}
-
-	let errorText = __('Failed to load', 'inser-core')
-	if (altText) {
-		errorText += ': ' + altText
-	}
-
-	const emptyElement = (
-		<Group
-			align="center"
-			position="center"
-			style={{
-				background: '#F8F9FA',
-				color: '#868E96',
-				height: height ?? 'auto',
-			}}
-		>
-			<IconCircleOff size={40} />
-			<Text size="xl" align="center">
-				{__('No image is set', 'inser-core')}
-			</Text>
-		</Group>
-	)
-
-	const imageElement = isUrlEmpty ? (
-		emptyElement
-	) : (
-		<>
-			<img
-				ref={imageRef}
-				onError={() => updateState({ hasError: true })}
-				src={imageUrl}
-				alt={altText}
-				style={{
-					width: '100%',
-					height: height ?? '100%',
-					objectFit: fit,
-				}}
-			/>
-			{hasError && (
-				<Group
-					align="center"
-					position="center"
-					style={{
-						inset: '0px',
-						position: 'absolute',
-						background: '#F8F9FA',
-						color: '#868E96',
-					}}
-				>
-					<IconPhotoOff size={40} />
-					<Text size="xl" align="center">
-						{errorText}
-					</Text>
-				</Group>
-			)}
-		</>
-	)
-
+export default function Edit(props: BlockEditProps<Attributes>) {
+	const { setAttributes, attributes } = props
 	return (
-		<Box>
-			<figure>
-				<div
-					style={{
-						height: height ?? 'auto',
-						position: 'relative',
-						textAlign: 'center',
-						...highlightBg,
-					}}
-				>
-					{renderResizable ? renderResizable(imageElement) : imageElement}
-				</div>
-
-				{!!caption && (
-					<Text component="figcaption" size="sm" align="center" color="dimmed">
-						{caption}
-					</Text>
-				)}
-			</figure>
-		</Box>
+		<SetupEditorEnv {...props} baseBlockName={'image'}>
+			<StateProvider stateCreator={storeCreator} keysToSave={Object.keys(json.attributes)} setAttributes={setAttributes} initialState={attributes}>
+				<EditComponent {...props} />
+			</StateProvider>
+		</SetupEditorEnv>
 	)
 }
