+++ src/blocks/image/index.tsx
@@ -1,13 +1,11 @@
 import { ConsumerBeacon } from '@inseri/lighthouse'
 import { IconPhoto } from '@tabler/icons-react'
 import { useBlockProps } from '@wordpress/block-editor'
-import type { BlockEditProps, BlockSaveProps } from '@wordpress/blocks'
+import type { BlockSaveProps } from '@wordpress/blocks'
 import { registerBlockType } from '@wordpress/blocks'
 import stringify from 'json-stable-stringify'
-import { SetupEditorEnv, StateProvider } from '../../components'
 import json from './block.json'
-import { PhotoEdit } from './Component'
-import { storeCreator } from './state'
+import Edit from './edit'
 
 const { name, ...settings } = json as any
 
@@ -21,28 +19,15 @@ export interface Attributes {
 	fit: 'cover' | 'contain' | 'fill' | 'scale-down' | 'none'
 }
 
-function Edit(props: BlockEditProps<Attributes>) {
-	const { setAttributes, attributes } = props
-	return (
-		<SetupEditorEnv {...props} baseBlockName={'image'}>
-			<StateProvider stateCreator={storeCreator} keysToSave={Object.keys(json.attributes)} setAttributes={setAttributes} initialState={attributes}>
-				<PhotoEdit {...props} />
-			</StateProvider>
-		</SetupEditorEnv>
-	)
-}
-
-function Save({ attributes }: BlockSaveProps<Attributes>) {
-	return (
-		<div {...useBlockProps.save()} data-attributes={stringify(attributes)}>
-			is loading ...
-		</div>
-	)
-}
-
 registerBlockType<Attributes>(name, {
 	...settings,
 	edit: Edit,
-	save: Save,
+	save: ({ attributes }: BlockSaveProps<Attributes>) => {
+		return (
+			<div {...useBlockProps.save()} data-attributes={stringify(attributes)}>
+				is loading ...
+			</div>
+		)
+	},
 	icon: <IconPhoto style={{ fill: 'none' }} />,
 })
