+++ src/blocks/localFileImport/view.tsx
@@ -1,41 +1,12 @@
-import { useControlTower, useDispatch } from '@inseri/lighthouse'
+import { useDispatch } from '@inseri/lighthouse'
 import { Dropzone } from '@mantine/dropzone'
-import { IconDownload, IconX, IconCheck } from '@tabler/icons-react'
-import { InspectorControls } from '@wordpress/block-editor'
-import type { BlockEditProps } from '@wordpress/blocks'
-import { PanelBody, PanelRow, ResizableBox, TextareaControl, TextControl, ToggleControl } from '@wordpress/components'
+import { IconCheck, IconDownload, IconX } from '@tabler/icons-react'
 import { useEffect } from '@wordpress/element'
-import { __ } from '@wordpress/i18n'
-import { ActionIcon, Button, createStyles, Group, MultiSelect, Stack, Text, useGlobalState, useMantineTheme } from '../../components'
-import { COMMON_CONTENT_TYPES, guessContentTypeByExtension, handleBody } from '../../utils'
-import { Attributes } from './index'
+import { ActionIcon, Button, Group, Stack, Text, createStyles, useGlobalState, useMantineTheme } from '../../components'
+import { guessContentTypeByExtension, handleBody } from '../../utils'
 import { GlobalState } from './state'
-import config from './block.json'
 
 const useStyles = createStyles((theme) => ({
-	label: {
-		textTransform: 'uppercase',
-		fontSize: '11px',
-		fontWeight: 500,
-		lineHeight: '1.4',
-		fontFamily: '-apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen-Sans, Ubuntu, Cantarell, "Helvetica Neue", sans-serif',
-		marginBottom: 'calc(8px)',
-	},
-	multiSelectValues: {
-		[`& > input[type="search"]`]: {
-			backgroundColor: 'unset',
-			minHeight: 'unset',
-			border: 0,
-		},
-
-		[`& > input[type="search"]:focus`]: {
-			backgroundColor: 'unset',
-			minHeight: 'unset',
-			border: 0,
-			boxShadow: 'unset',
-		},
-	},
-
 	itemButton: {
 		[`&:hover`]: {
 			backgroundColor: 'unset',
@@ -56,83 +27,13 @@ const useStyles = createStyles((theme) => ({
 	},
 }))
 
-const baseBeacon = { contentType: '', description: 'visitor provided data', key: 'data', default: '' }
-
-export function LocalFileImportEdit({ isSelected }: BlockEditProps<Attributes>) {
-	const { blockId, blockName, actions, accepts, mainText, subText, multiple, output } = useGlobalState((state: GlobalState) => state)
-	const { updateState } = actions
-	const { multiSelectValues, label } = useStyles().classes
-
-	const producersBeacons = useControlTower({ blockId, blockType: config.name, instanceName: blockName }, [baseBeacon])
-
-	useEffect(() => {
-		if (producersBeacons.length > 0 && !output.key) {
-			updateState({ output: producersBeacons[0] })
-		}
-	}, [producersBeacons.length])
-
-	const renderResizable = (children: JSX.Element) => (
-		<ResizableBox
-			showHandle={isSelected}
-			enable={{ bottom: true }}
-			minHeight={100}
-			onResize={(_event, _direction, element) => {
-				updateState({ height: element.offsetHeight })
-			}}
-		>
-			{children}
-		</ResizableBox>
-	)
-
-	return (
-		<>
-			<InspectorControls key="setting">
-				<PanelBody>
-					<PanelRow>
-						<TextControl label={__('Block Name', 'inseri-core')} value={blockName} onChange={(value) => updateState({ blockName: value })} />
-					</PanelRow>
-					<PanelRow>
-						<ToggleControl
-							label={__('Support multiple files', 'inseri-core')}
-							checked={multiple}
-							onChange={() => {
-								updateState({ multiple: !multiple })
-							}}
-						/>
-					</PanelRow>
-					<PanelRow>
-						<MultiSelect
-							styles={{ root: { width: '100%' } }}
-							classNames={{ values: multiSelectValues, label }}
-							data={COMMON_CONTENT_TYPES}
-							value={accepts}
-							onChange={(vals) => updateState({ accepts: vals })}
-							label={__('Which file types are accepted', 'inseri-core')}
-							mb="sm"
-							clearable
-							searchable
-						/>
-					</PanelRow>
-					<PanelRow>
-						<TextareaControl label={__('Primary Text', 'inseri-core')} value={mainText} onChange={(value) => updateState({ mainText: value })} />
-					</PanelRow>
-					<PanelRow>
-						<TextareaControl label={__('Secondary Text', 'inseri-core')} value={subText} onChange={(value) => updateState({ subText: value })} />
-					</PanelRow>
-				</PanelBody>
-			</InspectorControls>
-			<LocalFileImportView renderResizable={renderResizable} isGutenbergEditor isSelected={isSelected} />
-		</>
-	)
-}
-
 interface ViewProps {
 	isGutenbergEditor?: boolean
 	isSelected?: boolean
 	renderResizable?: (EditorComponent: JSX.Element) => JSX.Element
 }
 
-export function LocalFileImportView({ renderResizable, isGutenbergEditor, isSelected }: ViewProps) {
+export default function View({ renderResizable, isGutenbergEditor, isSelected }: ViewProps) {
 	const { accepts, mainText, subText, multiple, height, files, chosenFile, output } = useGlobalState((state: GlobalState) => state)
 	const { updateState, addFiles, removeFile } = useGlobalState((state: GlobalState) => state.actions)
