+++ src/blocks/mediaCollection/index.tsx
@@ -1,13 +1,11 @@
 import { IconFiles } from '@tabler/icons-react'
 import { useBlockProps } from '@wordpress/block-editor'
-import type { BlockEditProps, BlockSaveProps } from '@wordpress/blocks'
+import type { BlockSaveProps } from '@wordpress/blocks'
 import { registerBlockType } from '@wordpress/blocks'
 import stringify from 'json-stable-stringify'
-import { SetupEditorEnv, StateProvider } from '../../components'
 import { ProducerBeacon } from '../../globalScript'
 import json from './block.json'
-import { MediaLibraryEdit } from './Component'
-import { storeCreator } from './state'
+import Edit from './edit'
 
 const { name, ...settings } = json as any
 
@@ -21,28 +19,15 @@ export interface Attributes {
 	isVisible: boolean
 }
 
-function Edit(props: BlockEditProps<Attributes>) {
-	const { setAttributes, attributes } = props
-	return (
-		<SetupEditorEnv {...props} baseBlockName={'mediaCollection'}>
-			<StateProvider stateCreator={storeCreator} keysToSave={Object.keys(json.attributes)} setAttributes={setAttributes} initialState={attributes}>
-				<MediaLibraryEdit {...props} />
-			</StateProvider>
-		</SetupEditorEnv>
-	)
-}
-
-function Save({ attributes }: BlockSaveProps<Attributes>) {
-	return (
-		<div {...useBlockProps.save()} data-attributes={stringify(attributes)}>
-			is loading ...
-		</div>
-	)
-}
-
 registerBlockType<Attributes>(name, {
 	...settings,
 	edit: Edit,
-	save: Save,
+	save: ({ attributes }: BlockSaveProps<Attributes>) => {
+		return (
+			<div {...useBlockProps.save()} data-attributes={stringify(attributes)}>
+				is loading ...
+			</div>
+		)
+	},
 	icon: <IconFiles style={{ fill: 'none' }} />,
 })
