+++ src/blocks/python/block.json
@@ -2,7 +2,6 @@
 	"$schema": "https://schemas.wp.org/trunk/block.json",
 	"apiVersion": 2,
 	"name": "inseri-core/python",
-	"version": "0.1.0",
 	"title": "Python Code",
 	"category": "inseri",
 	"description": "Run Python code",
