+++ src/blocks/textViewer/block.json
@@ -2,7 +2,6 @@
 	"$schema": "https://schemas.wp.org/trunk/block.json",
 	"apiVersion": 2,
 	"name": "inseri-core/text-viewer",
-	"version": "0.1.0",
 	"title": "Text Viewer",
 	"category": "inseri",
 	"description": "View JSON, XML, HTML, ...",
