+++ src/blocks/webApi/block.json
@@ -2,7 +2,6 @@
 	"$schema": "https://schemas.wp.org/trunk/block.json",
 	"apiVersion": 2,
 	"name": "inseri-core/web-api",
-	"version": "0.1.0",
 	"title": "Web API",
 	"category": "inseri",
 	"description": "Load data from a Web API",
